refactor: jwt 디렉터리 이동
